Long -> Integer로 변경

+++ src/main/java/com/example/demo/src/scholarship_comment/model/GetScholarshipCommentRes.java
@@ -9,9 +9,9 @@ import lombok.Setter;
 @AllArgsConstructor
 public class GetScholarshipCommentRes {
 
-    private Long scholarship_comment_idx;
-    private Long scholarship_idx;
-    private Long user_idx;
+    private Integer scholarship_comment_idx;
+    private Integer scholarship_idx;
+    private Integer user_idx;
     private String scholarship_comment_content;
     private String scholarship_comment_updateAt;
