+++ src/main/java/com/example/demo/src/support_comment/model/PostSupportCommentRes.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostSupportCommentRes {
-    private Long support_comment_idx;
+    private Integer support_comment_idx;
 
 }
