+++ src/main/java/com/example/demo/src/support_comment/model/SupportComment.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class SupportComment {
-    private Long support_comment_idx;
-    private Long support_idx;
-    private Long user_idx;
+    private Integer support_comment_idx;
+    private Integer support_idx;
+    private Integer user_idx;
     private String support_comment_content;
 
 }
